# docs/api/openapi-skeleton.yaml
openapi: 3.1.0
info:
  title: Network CRM / Network OS API
  version: 0.1.0-draft
  description: >
    Draft skeleton only — endpoint groups and shared contracts, not final request/response
    schemas. Traces to docs/requirements/api-specification.md (API-001..024). Do not generate
    server code from this file until an ADR-reviewed team has filled in schemas per endpoint
    and resolved the QSTs referenced inline. All paths are tenant-scoped via the authenticated
    session (see docs/architecture/adr/0002-tenancy-and-storage.md); tenant_id is never a
    client-supplied path/query parameter used for authorization.
servers:
  - url: https://{tenant}.api.example.invalid/v1
    description: Placeholder host pattern — final domain/routing decided with ADR-0002/White Label design.
    variables:
      tenant:
        default: demo
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token issued by /auth (ADR-0003). Carries resolved tenant membership context.
  parameters:
...
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: Optimistic-locking ETag/version check for update operations.
      schema:
        type: string
...
  schemas:
    Error:
      type: object
      required: [code, message, correlation_id, retryable]
      properties:
        code:
          type: string
          description: See docs/api/error-catalog.md for the full enumerated list.
        message:
          type: string
          description: Safe, user-displayable message. Never includes PII or internal detail.
        field_errors:
          type: array
          items:
            type: object
            required: [field, code]
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
        correlation_id:
          type: string
          format: uuid
        retryable:
          type: boolean
...
  responses:
    Error:
      description: Standard error envelope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /roles:
    get:
      tags: [identity]
      responses:
        '200': { description: Role list }
    post:
      tags: [identity]
      summary: Create a scoped role (FR-ADMIN-001). Blocked if it would exceed the creator's own privileges.
      responses:
        '201': { description: Role created }
        default: { $ref: '#/components/responses/Error' }
  /roles/{id}/permissions:
    put:
      tags: [identity]
      summary: Assign scoped permissions to a role (privilege-escalation check, roles-and-permissions.md §4).
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Updated }
        default: { $ref: '#/components/responses/Error' }

  # ---- API-004/005 contacts ----
...